+++ TeamWork/MyGame/scr/scene/addScreen/GameOverScreen.cpp
@@ -35,7 +35,7 @@ void GameOverScreen::Init()
 {
 	inputCount_ = 0;
 	sinCount_ = defSinC;
-	for (int i = 0; i < changeSceneList_.size(); i++) {
+	for (int i = 0; i < (int)changeSceneList_.size(); i++) {
 		textAlphaList_[i] = 1.f;
 		textSizeList_[i] = 1.f;
 	}
@@ -120,11 +120,11 @@ void GameOverScreen::drawUpdate()
 	sinCount_+=3;
 	if (sinCount_ > 360)sinCount_ = 0;
 
-	for (int i = 0; i < changeSceneList_.size(); i++) {
+	for (int i = 0; i < (int)changeSceneList_.size(); i++) {
 		textAlphaList_[i]=1.f;
 		textSizeList_[i]=1.f;
 	}
-	textAlphaList_[inputCount_] = abs(MathHelper::Sin(sinCount_));
+	textAlphaList_[inputCount_] = abs(MathHelper::Sin((float)sinCount_));
 	textSizeList_[inputCount_] = mxmSize;
 
 }
